Update tournament system data structures and versioning
- Bump version to 1.0.73 in Directory.Build.props.
- Simplify leaderboard population in BasePlayerClient.cs.
- Change Leaderboard type in GameConsoleUI to List<LeaderboardDto>.
- Refactor BotTestRunner to use Match object for clarity.
- Update LeaderboardEntryTests to use uint for score counts.
- Modify TournamentTests to reflect new leaderboard structure.
- Add JSON serialization attributes to multiple model classes.
- Change Tournament leaderboard from dictionary to list.
- Update GameServer and TournamentManager to use byte for moves.
- Adjust front-end JavaScript to align with new leaderboard structure.
- Ensure HTML and Razor files display correct data and manage cache.
- Update AzureStorageService for new tournament state data structures.
- Overall refactor improves type safety, serialization, and code clarity.

+++ src/TicTacToe.Tournament.Server/Interfaces/IAzureStorageService.cs
@@ -10,10 +10,10 @@ public interface IAzureStorageService
     Task SaveTournamentStateAsync(TournamentContext tContext);
 
     Task<(
-        Models.Tournament? Tournament, 
-        List<PlayerInfo>? PlayerInfos, 
-        Dictionary<Guid, Guid>? Map, 
-        ConcurrentDictionary<Guid, ConcurrentQueue<(int Row, int Col)>>? Moves)>
+        Models.Tournament? Tournament,
+        List<PlayerInfo>? PlayerInfos,
+        Dictionary<Guid, Guid>? Map,
+        ConcurrentDictionary<Guid, ConcurrentQueue<(byte Row, byte Col)>>? Moves)>
         LoadTournamentStateAsync(Guid tournamentId);
 
     Task DeleteTournamentAsync(Guid tournamentId);
